Move CPUAffinity to utils
